refactor: change the location of module file
