+++ src/workoutLog/workoutLog.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { WorkoutLog } from '../domain/WorkoutLog.entity';
+import { WorkoutLog } from './domain/WorkoutLog.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([WorkoutLog])],
